fix: oop 구조 수정
