+++ src/02-controller/req.validator/curation/delete.curation.req.validator.js
