+++ src/02-controller/req.validator/curation/get.curation.list.req.validator.js
