+++ src/02-controller/req.validator/style/delete.style.req.validator.js
