+++ src/02-controller/req.validator/style/list.style.req.validator.js
@@ -1,0 +1,61 @@
+import { Exception, EXCEPTIONS } from "../../../common/exception.js";
+import { BaseValidator } from "../base.validator.js";
+
+export class ListStyleValidator extends BaseValidator {
+  constructor(data) {
+    super(data);
+  }
+
+  validate() {
+    const { page, pageSize, sortBy, tag, searchBy, keyword } = this.query;
+
+    const rawPage = this.isEmpty(page) ? "1" : page;
+    const rawPageSize = this.isEmpty(pageSize) ? "10" : pageSize;
+    const rawSortBy = this.isEmpty(sortBy) ? "latest" : sortBy;
+
+    if (!/^\d+$/.test(String(rawPage))) throw new Exception(EXCEPTIONS.PAGE_FORM);
+    const pageNum = Number(rawPage);
+    if (pageNum < 1) throw new Exception(EXCEPTIONS.PAGE_FORM);
+
+    if (!/^\d+$/.test(String(rawPageSize))) throw new Exception(EXCEPTIONS.PAGESIZE_FORM);
+    const pageSizeNum = Number(rawPageSize);
+    if (pageSizeNum < 1) throw new Exception(EXCEPTIONS.PAGESIZE_FORM);
+
+    const allowedSort = ["latest", "view", "curation"];
+    if (!this.isString(rawSortBy) || !allowedSort.includes(rawSortBy)) {
+      throw new Exception(EXCEPTIONS.BAD_REQUEST);
+    }
+
+    if (!this.isEmpty(tag) && !this.isString(tag)) {
+      throw new Exception(EXCEPTIONS.BAD_REQUEST);
+    }
+
+    let finalSearchBy = undefined;
+    let finalKeyword = undefined;
+    if (!this.isEmpty(searchBy) || !this.isEmpty(keyword)) {
+      if (this.isEmpty(searchBy)) throw new Exception(EXCEPTIONS.SEARCHBY_FORM);
+      if (this.isEmpty(keyword)) throw new Exception(EXCEPTIONS.KEYWORD_FORM);
+
+      const allowedSearch = ["nickname", "title", "content", "tag"];
+      if (!this.isString(searchBy) || !allowedSearch.includes(searchBy)) {
+        throw new Exception(EXCEPTIONS.SEARCHBY_FORM);
+      }
+
+      if (!this.isString(keyword) || keyword.length < 1 || keyword.length > 100) {
+        throw new Exception(EXCEPTIONS.KEYWORD_FORM);
+      }
+
+      finalSearchBy = searchBy;
+      finalKeyword = keyword;
+    }
+
+    return {
+      page: pageNum,
+      pageSize: pageSizeNum,
+      sortBy: rawSortBy,
+      tag: this.isEmpty(tag) ? undefined : tag,
+      searchBy: finalSearchBy,
+      keyword: finalKeyword,
+    };
+  }
+}
