+++ src/02-controller/res.dto/curation/delete.curation.res.dto.js
