+++ src/02-controller/res.dto/style/create.style.res.dto.js
