+++ src/02-controller/res.dto/style/delete.style.res.dto.js
