+++ src/02-controller/res.dto/style/get.style.detail.res.dto.js
