--- src/03-controller/curation.controller.js
@@ -1,29 +1,0 @@
-import { BaseController } from "./base.controller.js";
-
-export class CurationController extends BaseController {
-  #curationMiddleware;
-  constructor(curationMiddleware) {
-    super("");
-    this.#curationMiddleware = curationMiddleware;
-    this.registerRoutes();
-  }
-
-  registerRoutes = () => {
-    this.router.post(
-      "/styles/:styleId/curations",
-      this.catchException(this.#curationMiddleware.createCurationMiddleware),
-    );
-    this.router.get(
-      "/styles/:styleId/curations",
-      this.catchException(this.#curationMiddleware.getCurationListMiddleware),
-    );
-    this.router.put(
-      "/curations/:curationId",
-      this.catchException(this.#curationMiddleware.updateCurationMiddleware),
-    );
-    this.router.delete(
-      "/curations/:curationId",
-      this.catchException(this.#curationMiddleware.deleteCurationMiddleware),
-    );
-  };
-}
