--- src/03-controller/style.controller.js
@@ -1,38 +1,0 @@
-import { BaseController } from "./base.controller.js";
-
-export class StyleController extends BaseController {
-  #styleMiddleware;
-
-  constructor(styleMiddleware) {
-    super("");
-    this.#styleMiddleware = styleMiddleware;
-    this.registerRoutes();
-  }
-
-  registerRoutes() {
-    this.router.get(
-      "/styles",
-      this.catchException(this.#styleMiddleware.galleryListMiddleware),
-    );
-    this.router.get(
-      "/styles/popular-tags",
-      this.catchException(this.#styleMiddleware.popularTagsMiddleware),
-    );
-    this.router.post(
-      "/styles",
-      this.catchException(this.#styleMiddleware.createStyleMiddleware),
-    );
-    this.router.get(
-      "/styles/:styleId",
-      this.catchException(this.#styleMiddleware.getStyleDetailMiddleware),
-    );
-    this.router.put(
-      "/styles/:styleId",
-      this.catchException(this.#styleMiddleware.updateStyleMiddleware),
-    );
-    this.router.delete(
-      "/styles/:styleId",
-      this.catchException(this.#styleMiddleware.deleteStyleMiddleware),
-    );
-  }
-}
